pyproject.toml
[project]
name = "fastapi-ca"
version = "0.1.0"
description = ""
authors = [
    {name = "Jaekyoung Kim",email = "rlakim5521@naver.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "py-ulid (>=1.0.3,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "dependency-injector (>=4.48.2,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false